+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [6.5.19] - 
 ###Added
 - ParentId on start, init and error requests
+- Track total downloaded bytes 
 ## [6.5.18] - 2020-04-06
 ###Added
 - YBOptionUtilsKeys to easly share new props with config utils